requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Mail
Flask-WTF
Flask-Login
email-validator
PyJWT